src/auth/requirements.txt
boto3
python-dotenv
python-decouple